Make the *Help* buffer sticky in popwin.el
"Sticky" means that when you open the *Help* buffer and then move the
focus aways from it (e.g., `C-x C-o`) then it doesn't close, which it
does if it's not sticky. This is useful for the *Help* buffer so that I
can have documentation at the bottom while writing code.